bert_module/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers[torch] # Installiert PyTorch als Backend für Transformers
sentencepiece # Oft von BERT-Tokenizern benötigt